LPS-66031 Rename files for consistency
