+++ modules/apps/knowledge-base/knowledge-base-web/src/main/java/com/liferay/knowledge/base/web/portlet/configuration/icon/UnsubscribeKBArticlePortletConfigurationIcon.java
@@ -16,35 +16,44 @@ package com.liferay.knowledge.base.web.portlet.configuration.icon;
 
 import com.liferay.knowledge.base.constants.KBActionKeys;
 import com.liferay.knowledge.base.constants.KBPortletKeys;
-import com.liferay.knowledge.base.service.permission.AdminPermission;
+import com.liferay.knowledge.base.model.KBArticle;
+import com.liferay.knowledge.base.service.permission.KBArticlePermission;
+import com.liferay.knowledge.base.web.constants.KBWebKeys;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIcon;
 import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.service.SubscriptionLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 
+import javax.portlet.ActionRequest;
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
 import javax.portlet.PortletURL;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Sergio González
+ * @author Ambrin Chaudhary
  */
 @Component(
 	immediate = true,
-	property = {"javax.portlet.name=" + KBPortletKeys.KNOWLEDGE_BASE_ADMIN},
+	property = {
+		"javax.portlet.name=" + KBPortletKeys.KNOWLEDGE_BASE_ADMIN,
+		"path=/admin/view_article.jsp"
+	},
 	service = PortletConfigurationIcon.class
 )
-public class TemplatesPortletConfigurationIcon
+public class UnsubscribeKBArticlePortletConfigurationIcon
 	extends BasePortletConfigurationIcon {
 
 	@Override
 	public String getMessage(PortletRequest portletRequest) {
 		return LanguageUtil.get(
-			getResourceBundle(getLocale(portletRequest)), "templates");
+			getResourceBundle(getLocale(portletRequest)), "unsubscribe");
 	}
 
 	@Override
@@ -53,21 +62,30 @@ public class TemplatesPortletConfigurationIcon
 
 		PortletURL portletURL = PortalUtil.getControlPanelPortletURL(
 			portletRequest, KBPortletKeys.KNOWLEDGE_BASE_ADMIN,
-			PortletRequest.RENDER_PHASE);
+			PortletRequest.ACTION_PHASE);
 
-		portletURL.setParameter("mvcPath", "/admin/view_templates.jsp");
+		portletURL.setParameter(
+			ActionRequest.ACTION_NAME, "unsubscribeKBArticle");
 
 		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
 		portletURL.setParameter("redirect", themeDisplay.getURLCurrent());
 
+		KBArticle kbArticle = (KBArticle)portletRequest.getAttribute(
+			KBWebKeys.KNOWLEDGE_BASE_KB_ARTICLE);
+
+		portletURL.setParameter(
+			"resourceClassNameId", String.valueOf(kbArticle.getClassNameId()));
+		portletURL.setParameter(
+			"resourcePrimKey", String.valueOf(kbArticle.getResourcePrimKey()));
+
 		return portletURL.toString();
 	}
 
 	@Override
 	public double getWeight() {
-		return 101;
+		return 110;
 	}
 
 	@Override
@@ -75,10 +93,18 @@ public class TemplatesPortletConfigurationIcon
 		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		if (AdminPermission.contains(
-				themeDisplay.getPermissionChecker(),
-				themeDisplay.getScopeGroupId(),
-				KBActionKeys.VIEW_KB_TEMPLATES)) {
+		KBArticle kbArticle = (KBArticle)portletRequest.getAttribute(
+			KBWebKeys.KNOWLEDGE_BASE_KB_ARTICLE);
+
+		PermissionChecker permissionChecker =
+			themeDisplay.getPermissionChecker();
+
+		if ((kbArticle.isApproved() || !kbArticle.isFirstVersion()) &&
+			KBArticlePermission.contains(
+				permissionChecker, kbArticle, KBActionKeys.SUBSCRIBE) &&
+			_subscriptionLocalService.isSubscribed(
+				themeDisplay.getCompanyId(), themeDisplay.getUserId(),
+				KBArticle.class.getName(), kbArticle.getResourcePrimKey())) {
 
 			return true;
 		}
@@ -86,9 +112,13 @@ public class TemplatesPortletConfigurationIcon
 		return false;
 	}
 
-	@Override
-	public boolean isToolTip() {
-		return false;
+	@Reference(unbind = "-")
+	protected void setSubscriptionLocalService(
+		SubscriptionLocalService subscriptionLocalService) {
+
+		_subscriptionLocalService = subscriptionLocalService;
 	}
+
+	private SubscriptionLocalService _subscriptionLocalService;
 
 }
